change zip file

+++ src/GameObject/Plants/WallNut.cpp
@@ -11,7 +11,7 @@ void WallNut::Update()
         return;
     }
 
-    if (m_HP < 4000 / 3 && !m_is_cracked)
+    if (m_HP < 4000 / (double) 3 && !m_is_cracked)
     {
         m_is_cracked = true;
         ChangeImage(IMGID_WALLNUT_CRACKED);
